Change imports after moving findinfiles widgets.
